moved trainable kernel in techniques
